refactor(P4b-4): extract ToolInjection mixin verbatim from Executor
Native tool catalog assembly (special-pinned + caller-supplied + registry-
discovered + GAIA-triggered), passthrough policy gating, dedup variants,
and per-tier injection limits move into Inference::Executor::ToolInjection.
Verbatim move; lazy memos (@native_dispatch_tools, @native_tool_definitions,
@cached_injected_system) preserved exactly.

Executor.rb: 1681 → 1343 lines. tool_injection.rb: 356 lines.
Total executor.rb reduction across P4b-1..4: 2862 → 1343 (-1519, ~53%).
